Attempt implementation.
2002-06-27  Morten Welinder  <terra@diku.dk>

	* gsf/gsf-input-gzip.c (gsf_input_gzip_seek): Attempt
	implementation.

	* gsf/gsf-input.c (gsf_input_seek): Make pos an off_t.

+++ gsf/gsf-input.c
@@ -217,7 +217,7 @@ gsf_input_tell (GsfInput *input)
 gboolean
 gsf_input_seek (GsfInput *input, off_t offset, GsfOff_t whence)
 {
-	int pos = offset;
+	off_t pos = offset;
 
 	g_return_val_if_fail (input != NULL, -1);
